+++ src/components/parts/_legacy/main-article/MainArticleNewsPress.module.scss
